+++ crates/rome_js_analyze/tests/specs/nursery/noGlobalIsNan/invalid.js
@@ -1,0 +1,17 @@
+isNaN({});
+
+(isNaN)({});
+
+globalThis.isNaN({});
+
+(globalThis).isNaN({});
+
+globalThis["isNaN"]({});
+
+(globalThis)[("isNaN")]({});
+
+function localIsNaN(isNaN) {
+    globalThis.isNaN({});
+}
+
+localIsNaN(isNaN);