+++ crates/rome_js_analyze/tests/specs/nursery/noGlobalIsNan/valid.js
@@ -1,0 +1,14 @@
+Number.isNaN(Number.NaN);
+
+globalThis.Number.isNaN(Number.NaN);
+
+function localIsNaN(isNaN) {
+    isNaN({});
+}
+
+function localVar() {
+    var isNaN;
+    isNaN()
+}
+
+localIsNaN(Number.isNaN);